project/orchestrator/requirements.txt
paho-mqtt
configparser
requests